scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Tools for processing HumbleBundle books"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=10.0.0",
]
